Sprites
replace most of the sprites with ones that link to the theme. I also made a list of the sprites and how they link to the theme and also what should be collected and what shouldn't. I also made it so that the danger cells will spawn away from the player, so the player doesn't instantly get hit with one once they collect a book